Move authorisation functionality out of the application controller and into its own module
